# books.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "books.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/shelves": {
      "get": {
        "operationId": "ListShelves",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/ListShelvesResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "page_token",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "LibraryService"
        ]
      },
      "post": {
        "operationId": "CreateShelf",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/Shelf"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Shelf"
            }
          }
        ],
        "tags": [
          "LibraryService"
        ]
      }
    },
    "/v1/{book.name}": {
      "patch": {
        "operationId": "UpdateBook",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/Book"
            }
          }
        },
        "parameters": [
          {
            "name": "book.name",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Book"
            }
          }
        ],
        "tags": [
          "LibraryService"
        ]
      }
    },
    "/v1/{name}": {
      "get": {
        "operationId": "GetBook",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/Book"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "LibraryService"
        ]
      },
      "delete": {
        "operationId": "DeleteBook",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/protobufEmpty"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "LibraryService"
        ]
      }
    },
    "/v1/{parent}/books": {
      "get": {
        "operationId": "ListBooks",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/ListBooksResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "parent",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "page_size",
            "description": "The maximum number of items to return.",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "page_token",
            "description": "The next_page_token value returned from a previous List request, if any.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "LibraryService"
        ]
      },
      "post": {
        "operationId": "CreateBook",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/Book"
            }
          }
        },
        "parameters": [
          {
            "name": "parent",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Book"
            }
          }
        ],
        "tags": [
          "LibraryService"
        ]
      }
    }
  },
  "definitions": {
    "Book": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "description": "Resource name of the book. It must have the format of \"shelves/*/books/*\"."
        },
        "text": {
          "type": "string",
          "title": "... other properties"
        }
      }
    },
    "CreateBookRequest": {
      "type": "object",
      "properties": {
        "parent": {
          "type": "string",
          "description": "The parent resource name where the book is to be created."
        },
        "book_id": {
          "type": "string",
          "description": "The book id to use for this book."
        },
        "book": {
          "$ref": "#/definitions/Book",
          "description": "The book resource to create.\nThe field name should match the Noun in the method name."
        }
      }
    },
    "CreateShelfRequest": {
      "type": "object",
      "properties": {
        "shelf": {
          "$ref": "#/definitions/Shelf"
        }
      }
    },
    "DeleteBookRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "The resource name of the book to be deleted, for example:\n\"shelves/shelf1/books/book2\""
        }
      }
    },
    "DeleteShelfRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
    "GetBookRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "The field will contain name of the resource requested, for example:\n\"shelves/shelf1/books/book2\""
        }
      }
    },
    "ListBooksRequest": {
      "type": "object",
      "properties": {
        "parent": {
          "type": "string",
          "description": "The parent resource name, for example, \"shelves/shelf1\"."
        },
        "page_size": {
          "type": "integer",
          "format": "int32",
          "description": "The maximum number of items to return."
        },
        "page_token": {
          "type": "string",
          "description": "The next_page_token value returned from a previous List request, if any."
        }
      }
    },
    "ListBooksResponse": {
      "type": "object",
      "properties": {
        "books": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Book"
          },
          "description": "The field name should match the noun \"books\" in the method name. There\nwill be a maximum number of items returned based on the page_size field\nin the request."
        },
        "next_page_token": {
          "type": "string",
          "description": "Token to retrieve the next page of results, or empty if there are no\nmore results in the list."
        }
      }
    },
    "ListShelvesRequest": {
      "type": "object",
      "properties": {
        "page_size": {
          "type": "integer",
          "format": "int32"
        },
        "page_token": {
          "type": "string"
        }
      }
    },
    "ListShelvesResponse": {
      "type": "object",
      "properties": {
        "shelves": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Shelf"
          }
        },
        "next_page_token": {
          "type": "string"
        }
      }
    },
    "Shelf": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
    "UpdateBookRequest": {
      "type": "object",
      "properties": {
        "book": {
          "$ref": "#/definitions/Book",
          "description": "The book resource which replaces the resource on the server."
        },
        "update_mask": {
          "$ref": "#/definitions/protobufFieldMask",
          "description": "The update mask applies to the resource."
        }
      }
    },
    "protobufEmpty": {
      "type": "object",
      "description": "service Foo {\n      rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty);\n    }\n\nThe JSON representation for `Empty` is empty JSON object `{}`.",
      "title": "A generic empty message that you can re-use to avoid defining duplicated\nempty messages in your APIs. A typical example is to use it as the request\nor the response type of an API method. For instance:"
    },
    "protobufFieldMask": {
      "type": "object",
      "properties": {
        "paths": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "description": "The set of field mask paths."
        }
      },
      "description": "paths: \"f.a\"\n    paths: \"f.b.d\"\n\nHere `f` represents a field in some root message, `a` and `b`\nfields in the message found in `f`, and `d` a field found in the\nmessage in `f.b`.\n\nField masks are used to specify a subset of fields that should be\nreturned by a get operation or modified by an update operation.\nField masks also have a custom JSON encoding (see below).\n\n# Field Masks in Projections\n\nWhen used in the context of a projection, a response message or\nsub-message is filtered by the API to only contain those fields as\nspecified in the mask. For example, if the mask in the previous\nexample is applied to a response message as follows:\n\n    f {\n      a : 22\n      b {\n        d : 1\n        x : 2\n      }\n      y : 13\n    }\n    z: 8\n\nThe result will not contain specific values for fields x,y and z\n(their value will be set to the default, and omitted in proto text\noutput):\n\n\n    f {\n      a : 22\n      b {\n        d : 1\n      }\n    }\n\nA repeated field is not allowed except at the last position of a\npaths string.\n\nIf a FieldMask object is not present in a get operation, the\noperation applies to all fields (as if a FieldMask of all fields\nhad been specified).\n\nNote that a field mask does not necessarily apply to the\ntop-level response message. In case of a REST get operation, the\nfield mask applies directly to the response, but in case of a REST\nlist operation, the mask instead applies to each individual message\nin the returned resource list. In case of a REST custom method,\nother definitions may be used. Where the mask applies will be\nclearly documented together with its declaration in the API.  In\nany case, the effect on the returned resource/resources is required\nbehavior for APIs.\n\n# Field Masks in Update Operations\n\nA field mask in update operations specifies which fields of the\ntargeted resource are going to be updated. The API is required\nto only change the values of the fields as specified in the mask\nand leave the others untouched. If a resource is passed in to\ndescribe the updated values, the API ignores the values of all\nfields not covered by the mask.\n\nIf a repeated field is specified for an update operation, the existing\nrepeated values in the target resource will be overwritten by the new values.\nNote that a repeated field is only allowed in the last position of a `paths`\nstring.\n\nIf a sub-message is specified in the last position of the field mask for an\nupdate operation, then the existing sub-message in the target resource is\noverwritten. Given the target message:\n\n    f {\n      b {\n        d : 1\n        x : 2\n      }\n      c : 1\n    }\n\nAnd an update message:\n\n    f {\n      b {\n        d : 10\n      }\n    }\n\nthen if the field mask is:\n\n paths: \"f.b\"\n\nthen the result will be:\n\n    f {\n      b {\n        d : 10\n      }\n      c : 1\n    }\n\nHowever, if the update mask was:\n\n paths: \"f.b.d\"\n\nthen the result would be:\n\n    f {\n      b {\n        d : 10\n        x : 2\n      }\n      c : 1\n    }\n\nIn order to reset a field's value to the default, the field must\nbe in the mask and set to the default value in the provided resource.\nHence, in order to reset all fields of a resource, provide a default\ninstance of the resource and set all fields in the mask, or do\nnot provide a mask as described below.\n\nIf a field mask is not present on update, the operation applies to\nall fields (as if a field mask of all fields has been specified).\nNote that in the presence of schema evolution, this may mean that\nfields the client does not know and has therefore not filled into\nthe request will be reset to their default. If this is unwanted\nbehavior, a specific service may require a client to always specify\na field mask, producing an error if not.\n\nAs with get operations, the location of the resource which\ndescribes the updated values in the request message depends on the\noperation kind. In any case, the effect of the field mask is\nrequired to be honored by the API.\n\n## Considerations for HTTP REST\n\nThe HTTP kind of an update operation which uses a field mask must\nbe set to PATCH instead of PUT in order to satisfy HTTP semantics\n(PUT must only be used for full updates).\n\n# JSON Encoding of Field Masks\n\nIn JSON, a field mask is encoded as a single string where paths are\nseparated by a comma. Fields name in each path are converted\nto/from lower-camel naming conventions.\n\nAs an example, consider the following message declarations:\n\n    message Profile {\n      User user = 1;\n      Photo photo = 2;\n    }\n    message User {\n      string display_name = 1;\n      string address = 2;\n    }\n\nIn proto a field mask for `Profile` may look as such:\n\n    mask {\n      paths: \"user.display_name\"\n      paths: \"photo\"\n    }\n\nIn JSON, the same mask is represented as below:\n\n    {\n      mask: \"user.displayName,photo\"\n    }\n\n# Field Masks and Oneof Fields\n\nField masks treat fields in oneofs just as regular fields. Consider the\nfollowing message:\n\n    message SampleMessage {\n      oneof test_oneof {\n        string name = 4;\n        SubMessage sub_message = 9;\n      }\n    }\n\nThe field mask can be:\n\n    mask {\n      paths: \"name\"\n    }\n\nOr:\n\n    mask {\n      paths: \"sub_message\"\n    }\n\nNote that oneof type names (\"test_oneof\" in this case) cannot be used in\npaths.",
      "title": "`FieldMask` represents a set of symbolic field paths, for example:"
    }
  }
}